feature: add sign in with Apple quality: extract all models to each models files - Code improvement - Rename files, methods, Viewmodel - extract viewmodel from map view to new file
