+++ SeaSwiftAndSun-Simplon/SeaSwiftAndSun-Simplon/AppDelegate.swift
@@ -6,16 +6,15 @@
 //
 
 import UIKit
+import FirebaseCore
 
 @main
 class AppDelegate: UIResponder, UIApplicationDelegate {
 
-
-
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
-        // Override point for customization after application launch.
+        FirebaseApp.configure()
         return true
-    }
+      }
 
     // MARK: UISceneSession Lifecycle
