+++ SeaSwiftAndSun-Simplon/SeaSwiftAndSun-Simplon/Controllers/ViewController.swift
@@ -6,6 +6,7 @@
 //
 
 import UIKit
+import SwiftUI
 
 class ViewController: UIViewController {
