+++ SeaSwiftAndSun-Simplon/SeaSwiftAndSun-Simplon/GoogleService-Info.plist
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>API_KEY</key>
+	<string>AIzaSyBtwC1TrKTUErgoa0epNmbYwz3BVyJnmNQ</string>
+	<key>GCM_SENDER_ID</key>
+	<string>607300669970</string>
+	<key>PLIST_VERSION</key>
+	<string>1</string>
+	<key>BUNDLE_ID</key>
+	<string>com.mahdia.SeaTravel</string>
+	<key>PROJECT_ID</key>
+	<string>signinwithapple-3cc19</string>
+	<key>STORAGE_BUCKET</key>
+	<string>signinwithapple-3cc19.appspot.com</string>
+	<key>IS_ADS_ENABLED</key>
+	<false></false>
+	<key>IS_ANALYTICS_ENABLED</key>
+	<false></false>
+	<key>IS_APPINVITE_ENABLED</key>
+	<true></true>
+	<key>IS_GCM_ENABLED</key>
+	<true></true>
+	<key>IS_SIGNIN_ENABLED</key>
+	<true></true>
+	<key>GOOGLE_APP_ID</key>
+	<string>1:607300669970:ios:cc262a84b46cddd40b3ecb</string>
+</dict>
+</plist>
