+++ SeaSwiftAndSun-Simplon/SeaSwiftAndSun-Simplon/Model/Full.swift
@@ -1,0 +1,14 @@
+//
+//  Full.swift
+//  SeaSwiftAndSun-Simplon
+//
+//  Created by Jonathan Duong on 14/12/2023.
+//
+
+import Foundation
+
+struct Full: Codable {
+    var url: String?
+    var width: Int?
+    var height: Int?
+}
