+++ SeaSwiftAndSun-Simplon/SeaSwiftAndSun-Simplon/Model/Record.swift
@@ -1,0 +1,14 @@
+//
+//  Record.swift
+//  SeaSwiftAndSun-Simplon
+//
+//  Created by Jonathan Duong on 11/12/2023.
+//
+
+import Foundation
+
+struct Record: Codable {
+    var id: String?
+    var createdTime: String?
+    var fields: Fields?
+}
