+++ SeaSwiftAndSun-Simplon/SeaSwiftAndSun-Simplon/Model/Records.swift
@@ -1,0 +1,12 @@
+//
+//  Records.swift
+//  SeaSwiftAndSun-Simplon
+//
+//  Created by Jonathan Duong on 14/12/2023.
+//
+
+import Foundation
+
+struct Records: Codable {
+    var records: [Record]?
+}
