+++ SeaSwiftAndSun-Simplon/SeaSwiftAndSun-Simplon/Model/Thumbnails.swift
@@ -1,0 +1,14 @@
+//
+//  Thumbnails.swift
+//  SeaSwiftAndSun-Simplon
+//
+//  Created by Jonathan Duong on 14/12/2023.
+//
+
+import Foundation
+
+struct Thumbnails: Codable {
+    var small: Full?
+    var large: Full?
+    var full: Full?
+}
